refactor: (#47) WebSocket 에러 처리 구조 개선
- WsError 클래스를 wsError() 유틸 함수로 변경
- ErrorResponseBuilder static 메서드를 buildErrorResponse() 함수로 변경
- ROOM_ERROR_CODES 상수 사용으로 하드코딩 제거
