+++ src/common/filters/socket-exception.filter.ts
@@ -1,19 +1,17 @@
-import { ArgumentsHost, Catch, Logger } from '@nestjs/common';
+import { ArgumentsHost, Catch } from '@nestjs/common';
 import { BaseWsExceptionFilter, WsException } from '@nestjs/websockets';
 import { Socket } from 'socket.io';
-import { ErrorResponseBuilder } from '../builders/error-response.builder';
+import { buildErrorResponse } from '../utils/error-response.util';
 
 @Catch()
 export class SocketExceptionFilter extends BaseWsExceptionFilter {
-  private readonly logger = new Logger(SocketExceptionFilter.name);
-
   override catch(exception: unknown, host: ArgumentsHost) {
     const client = host.switchToWs().getClient<Socket>();
 
     const raw =
       exception instanceof WsException ? (exception.getError?.() ?? exception) : exception;
 
-    const normalized = ErrorResponseBuilder.build(raw, 400);
+    const normalized = buildErrorResponse(raw, 500);
 
     client.emit('exception', normalized);
   }
