+++ src/rooms/rooms.gateway.ts
@@ -6,7 +6,8 @@ import { IRoomsEventPublisher } from './events/rooms-event.publisher';
 import { WebSocketGateway, WebSocketServer } from '@nestjs/websockets';
 import { SocketExceptionFilter } from 'src/common/filters/socket-exception.filter';
 import { Logger, UseFilters } from '@nestjs/common';
-import { WsError } from 'src/common/utils/ws-error.util';
+import { ROOM_ERROR_CODES } from './constants/room.constant';
+import { wsError } from 'src/common/utils/ws-error.util';
 
 @UseFilters(SocketExceptionFilter)
 @WebSocketGateway({
@@ -31,7 +32,7 @@ export class RoomsGateway implements IRoomsEventPublisher {
     } catch (err) {
       this.logger.error(`Failed to emit event: ${event}`, err as Error);
 
-      throw WsError.internalServerError('ROOM_EVENT_FAILED');
+      throw wsError(400, ROOM_ERROR_CODES.ROOM_EVENT_FAILED);
     }
   }
